Implement JFR ID natives
This PR implements the type ID support for threads, classes and
stacktraces. This IDs are used in the JCL eventwriter to write JFR
events. Similar to how the JVM uses it in the native event writer.

Related: https://github.com/eclipse-openj9/openj9/issues/23764
Related: https://github.com/eclipse-openj9/openj9/issues/23775
Related: https://github.com/eclipse-openj9/openj9/issues/24372

+++ runtime/jcl/common/jdk_jfr_internal_JVM_common.cpp
@@ -204,7 +204,7 @@ Java_jdk_jfr_internal_JVM_getClassId(JNIEnv *env, jclass clazz, jclass targetCla
 	vmFuncs->internalEnterVMFromJNI(currentThread);
 	j9object_t classObject = J9_JNI_UNWRAP_REFERENCE(targetClass);
 	J9Class *targetClazz = J9OBJECT_CLAZZ(currentThread, classObject);
-	classID = (jlong)targetClazz->classID;
+	classID = targetClazz->classID;
 	vmFuncs->internalExitVMToJNI(currentThread);
 
 	return classID;
